+++ src/Models/UsersModel.php
@@ -233,19 +233,11 @@ class UsersModel extends RbacModel
      */
     protected function onRead(array $fields): array
     {
-        $fields = $this->transform($fields, [
+        return $this->transform($fields, [
             'meta' => [$this, 'jsonDecode'],
             'admin' => [$this, 'boolean'],
             'enabled' => [$this, 'boolean']
         ]);
-
-        if (isset($fields['meta'])) {
-            $meta = Arr::dot($fields['meta']);
-            ksort($meta);
-            $fields['meta'] = Arr::undot($meta);
-        }
-
-        return $fields;
     }
 
     /**
@@ -278,7 +270,7 @@ class UsersModel extends RbacModel
         }
 
         if (isset($fields['meta']) && is_array($fields['meta'])) {
-            $fields['meta'] = $this->updateNullableJsonField($this->ormService, $this->table_name, $this->primary_key, $existing->getPrimaryKey(), $this->getNullableJsonField(), $fields['meta']);
+            $fields['meta'] = $this->updateNullableJsonField($this->ormService, $this->table_name, $this->primary_key, $existing->getPrimaryKey(), 'meta', $fields['meta']);
         }
 
         return $fields;
@@ -407,16 +399,6 @@ class UsersModel extends RbacModel
      * | Traits
      * |--------------------------------------------------------------------------
      */
-
-    /**
-     * Trait: HasNullableJsonField
-     *
-     * @inheritDoc
-     */
-    public function getNullableJsonField(): string
-    {
-        return 'meta';
-    }
 
     /**
      * Trait: HasOmittedFields